ItemStack isSimiliar API to skip durability and name checks
Skip Durability is useful if you simply want to see if an item type is the same for weapons
Skip Item Name is useful if you want to treat a renamed item the same as an unrenamed item, when lore is involved

For example, lore can be used to identify a custom item, but then if a player renames it, the isSimilar breaks.

This new boolean allows you to verify if the item is the same ignoring the name field by temporarily nulling it during the check.